add potion_str_escape; refactor inspect to use it
Added a potion_str_escape method which returns an escaped version
of a string. This should respect utf8 as well.

potion_str_inspect now printfs that instead of the raw string.
Maybe it should instead call print on it.

It's a little hard to test the utf8 handling right now because of ragel
issues, so be wary.