add preliminary condition system
A preliminary condition-signalling system, for Lisp-style conditions and
condition handlers. VM only at this point (i.e. no JIT implementation).

Handler tables are stored in the stack, and created lazily. Signaling a
condition walks up the stack and finds the first handler that matches
the signal. Handlers are called in the context of the signaller. The
return value of the handler is returned to the signaller.

Because handlers are created lazily, this should have a negligible
effect on normal stack winding & unwinding. (Just one extra stack slot
required per call.) Signaling a condition will take O(size of stack).

Signals can be any Potion object; handlers must be a prototype.

Still TODO:
- Continue walking up the stack if a handler returns PN_NIL. (How the
  hell am I going to do this?)
- Signal conditions from the VM for undefined methods, etc.
- JIT it.
- Make a default handler (probably as part of the stdlib).
- Think about what happens when you signal inside a signal handler.

* core/internal.c: add PN_signal and PN_handle keywords
* core/compile.c: add bytecode generation for signal and handle
* core/vm.c: implement signal and handle opcodes