make condition handling backoffs work
Condition handlers can now return false to signify they don't want to
handle the condition. In this case, a condition handler higher on the
stack should be called.

The SIGNAL opcode is now replaced with a FIND_HANDLER opcode. Signalling
a condition condition is compiled into a loop that repeatedly calls
FIND_HANDLER to find the next handler, calls it, and repeats if the
return value is false. Finding the next matching handler in the stack
has been refactored into a potion_find_handler() function.

The handler table position on the stack has been moved to the very start
of the current frame, in order to make it possible for
potion_find_handler() to determine the top of the frame, given a handler
offset.

TODO:
 - Decide what to do when a condition propagates all the way to the top
   of the stack. Currently it just segfaults.
 - Make it possible for the VM to signal a condition, e.g. when a method
   lookup fails.
 - Write JIT version.